requirements.txt
# ====================================
# AIGC-Identification-Toolkit Requirements
# ====================================

# ============== 核心深度学习框架 ==============
torch>=2.4.0,<3.0
torchvision>=0.19.0
torchaudio>=2.4.0
transformers>=4.44.0
diffusers>=0.34.0 
accelerate>=0.33.0

# ============== 数值计算和科学计算 ==============
numpy>=1.24.0
scipy>=1.10.0
scikit-learn>=1.0.0
pandas>=2.0.0

# ============== 图像处理 ==============
Pillow>=10.4.0
opencv-python>=4.5.0
scikit-image>=0.19.0
imageio>=2.0.0
PyWavelets>=1.3.0
timm==0.9.16
lpips>=0.1.4
pytorch_msssim>=0.2.0
pycocotools>=2.0.0

# ============== 音频处理（AudioSeal核心） ==============
julius>=0.2.7
soundfile>=0.12.0
librosa>=0.10.0
audioread>=3.0.0
soxr>=0.3.0

# ============== 视频处理（VideoSeal核心） ==============
ffmpeg-python>=0.2.0
av>=12.0.0

# ============== 文本处理和NLP（PostMark核心） ==============
nltk>=3.7
spacy>=3.4.0
tokenizers>=0.19.0

# ============== 配置和工具 ==============
omegaconf>=2.0.0
PyYAML>=6.0
tqdm>=4.64.0
requests>=2.25.0
psutil>=5.8.0
click>=8.0.0

# ============== 模型管理 ==============
huggingface-hub>=0.20.0
safetensors>=0.4.0
datasets>=2.0.0

# ============== 深度学习工具 ==============
einops>=0.8.0

# ============== 可视化和监控 ==============
matplotlib>=3.5.0
tensorboard>=2.13.0
calflops>=0.3.0

# ============== 测试工具 ==============
pytest>=6.0

